Non-isolated namespace, app initialization
For the Concerto Plugin Architecture, we structure the engine with a
non-isolated namespace (it acts like its own rack app). We also use a
hook from the Conerto Plugin model which allows us to request to be
mounted and gives us an oppurtunity to set up configuration options.
Eventually, we will also have UI hooks.