Switch to isolated namespace engine style
This has some impact on linking between the main app and the plugin, and
vice versa. It is intended to prevent namespace clashes and also make
features like migration installation easier in the future.

For better consistency and reproducibility, some changes are included in
this commit to make concerto-hardware a little more like what is
generated by `rails plugin new --mountable`.

This will only work with Concerto at 4dca556 or better.