Remove migration generator
Rails now includes a feature to autmatically copy migrations from
engines when they are installed, so a generator to do it is no longer
needed. Normally, this new feature is accessed with

  rake railties:install:migrations

However, to make installation even simpler, recent versions of Concerto
(ec0059220f and up) will install any migrations that are in the engine's
db/migrate path but have not yet been copied to the main application on
boot. So to install the migration all you need to do is add the Gem and
restart the server. The migration will be installed and run.